requirements.txt
Django==4.0.3
django-macaddress==1.8.0
fontawesome-free==5.15.4
dbus-python==1.2.18
